+++ src/extensions/rules/rulesext.cpp
@@ -104,6 +104,7 @@ RulesClassExtension::RulesClassExtension(const RulesClass* this_ptr) :
     PausedRepairsFrame(6),
     EscortRange(-1),
     AbandonTargetEscortRange(-1),
+	IsFreeRadarOnLowPower(false),
 	BridgeArmor(ARMOR_NULL)
 {
     /**
@@ -261,6 +262,7 @@ void RulesClassExtension::Object_CRC(CRCEngine &crc) const
     crc(PausedRepairsFrame);
     crc(EscortRange);
     crc(AbandonTargetEscortRange);
+    crc(IsFreeRadarOnLowPower);
     crc(BridgeArmor);
 }
 
@@ -694,6 +696,7 @@ bool RulesClassExtension::General(CCINIClass &ini)
     PausedRepairsFrame = ini.Get_Int(GENERAL, "PausedRepairsFrame", PausedRepairsFrame);
     EscortRange = ini.Get_Lepton(GENERAL, "EscortRange", EscortRange);
     AbandonTargetEscortRange = ini.Get_Lepton(GENERAL, "AbandonTargetEscortRange", AbandonTargetEscortRange);
+    IsFreeRadarOnLowPower = ini.Get_Bool(GENERAL, "FreeRadarOnLowPower", IsFreeRadarOnLowPower);
 
     /**
      *  Allow replacing any signle movement zone with a copy of RA2's water MZone.