+++ src/extensions/rules/rulesext.h
@@ -270,6 +270,11 @@ public:
      *  before abandoning it and going back to escort its guard target.
      */
     int AbandonTargetEscortRange;
+	
+	/**
+	 *  Determines whether Free Radar would still apply even during Low Power
+	 */
+	int IsFreeRadarOnLowPower;
 
 	/*
     * The armor type used by bridges for damage calculation.